refactor(test): split StorageEngineBackedEventStoreTestSuite into DCB vs Aggregate hierarchy
The OverrideAppendCondition integration tests use DCB-specific patterns
(ORIGIN-based uniqueness, criteria narrowing by event type, conflict bypass)
that are not supported by aggregate-based storage engines. The aggregate-based
JPA engine uses sequence-number conflict detection via unique constraints,
ignoring AppendCondition criteria entirely.

Introduce a test suite hierarchy:

  StorageEngineBackedEventStoreTestSuite        (common: sourcing, streaming, tokens, conflicts)
  ├── DcbBasedStorageEngineBackedEventStoreTestSuite   (DCB: overrideAppendCondition tests)
  │   ├── InMemoryStorageEngineBackedEventStoreTest
  │   └── AxonServerStorageEngineBackedEventStoreIT
  └── AggregateBasedStorageEngineBackedEventStoreTestSuite  (aggregate: placeholder)
      └── AggregateBasedJpaStorageEngineBackedEventStoreIT

The 3 OverrideAppendCondition tests and DCB event records (CourseCreated,
StudentSubscribedToCourse, StudentUnsubscribedFromCourse) move from the base
suite to DcbBasedStorageEngineBackedEventStoreTestSuite. The base suite
exposes eventStore, awaitLatch, and RESOLVER as protected for subclass access.
